fix(viewer): clear the geomTier pin from the URL too, and split the two oversized modules
Addresses all three review findings on #2545.

P2, and a real bug in the first commit: `clearGeomTierOverride` removed only the
localStorage entry. `getGeomTierOverride` re-reads the query parameter on every
call and re-persists it, so on the originating `?geomTier=low` link the very next
load silently restored the pin -- and that link is precisely the case the Clear
action exists for. It now strips `geomTier` from the address bar via
`history.replaceState`, preserving any other parameters, which also stops the
user re-sharing a URL that re-pins the next reader.

My earlier browser check missed this because it reloaded the BARE url after
clicking Clear rather than the originating one. Re-verified on the right
scenario: on `?geomTier=low&foo=bar`, Clear leaves `?foo=bar`, and both an
in-place model reload and a full page reload now log `tier=medium`.

P1 x2, module size: rather than deepen two files already past the ~400-line
limit, the cohesive blocks are extracted.

- `store/geometryFidelity.ts` (197) takes the mode, the tier thresholds, the
  sticky override and the resolution rule. `constants.ts` re-exports the surface
  so its many importers keep one import site: 568 -> 485, i.e. below where this
  PR found it.
- `slices/geometryLoadSettings.ts` (153) takes the load-time settings state and
  actions -- merge-layers, fidelity mode, tier pin -- which share a contract none
  of the slice's other state has: read only on the NEXT load, folded into the
  geometry cache key, so each must prompt a reload. `uiSlice.ts` composes it, so
  the store shape is unchanged: 464 -> 401. `isModelLoaded` is injected rather
  than read through a cross-slice `get()`, so the new module needs no knowledge
  of the rest of the store and there is no import cycle.

The URL-strip test drives a `history.replaceState` stub that rewrites
`location.search` the way a browser does, so it asserts what the next
`getGeomTierOverride()` actually sees rather than that replaceState was called;
removing the strip turns it red.
